Base template Tests
