+++ src/templates/Base/Base.test.jsx
@@ -2,10 +2,11 @@ import { screen } from "@testing-library/react";
 import { describe, expect } from "@jest/globals";
 import { renderTheme } from "../../styles/render-theme";
 import { Base } from ".";
+import { mockBase } from "./mock";
 
 describe("<Base />", () => {
-  it("should render the component", () => {
-    renderTheme(<Base>Children</Base>);
-    expect(screen.getByRole("heading")).toBeInTheDocument();
+  it("should render the template Base", () => {
+    const { container } = renderTheme(<Base {...mockBase} />);
+    expect(container).toMatchSnapshot();
   });
 });
