+++ src/templates/Base/index.jsx
@@ -4,7 +4,7 @@ import { Menu } from "../../components/Menu/";
 import { Footer } from "../../components/Footer";
 import { GoTop } from "../../components/GoTop";
 
-export const Base = ({ links = [], logoData, footerHtml, children }) => {
+export const Base = ({ links, logoData, footerHtml, children }) => {
   return (
     <>
       <Menu links={links} logoData={logoData} />
