--- src/templates/Base/mock.js
@@ -1,11 +1,0 @@
-import mock from "../../components/NavLinks/mock";
-import { NavLinks } from "../../components/NavLinks/index";
-
-export default {
-  links: mock,
-  logoData: {
-    text: "Logo",
-    link: "#",
-  },
-  footerHtml: `<p>Teste de Footer</a></p>`,
-};
