+++ src/templates/Base/stories.jsx
@@ -1,22 +1,8 @@
 import { Base } from ".";
 
-import mock from "./mock";
 import { GridText } from "../../components/GridText";
-import gridMock from "../../components/GridText/mock";
-
-export const mockBase = {
-  children: (
-    <>
-      <GridText {...gridMock} background />
-      <GridText {...gridMock} />
-      <GridText {...gridMock} background />
-      <GridText {...gridMock} />
-      <GridText {...gridMock} background />
-      <GridText {...gridMock} />
-    </>
-  ),
-  ...mock,
-};
+import mock from "./mock";
+import mockBase from "./mock";
 
 export default {
   title: "Templates/Base",
