+++ src/templates/Home/Home.test.jsx
@@ -2,17 +2,12 @@ import React from "react";
 import { screen } from "@testing-library/react";
 import { describe, expect, it } from "@jest/globals";
 import { renderTheme } from "../../styles/render-theme";
+import { theme } from "../../styles/theme";
 import Home from ".";
 
 describe("<Home />", () => {
   it("should render", () => {
-    renderTheme(<Home />);
-    expect(screen.getByRole("heading")).toBeInTheDocument();
+    const { container } = renderTheme(<Home />);
+    expect(container).toMatchSnapshot();
   });
-});
-
-it("renders learn react link", () => {
-  const { debug } = renderTheme(<Home />);
-
-  debug();
 });
